Fixes int overflow in Zxcvbn/DBScorer when calculating password weakness

+++ Zxcvbn/DBScorer.m
@@ -366,8 +366,8 @@ static int kNumDays = 31;
     int hour = minute * 60;
     int day = hour * 24;
     int month = day * 31;
-    int year = month * 12;
-    int century = year * 100;
+    NSInteger year = month * 12;
+    NSInteger century = year * 100;
     if (seconds < minute)
         return @"instant";
     if (seconds < hour)
